Separate out drag and drop logic into component
