+++ src/App.js
@@ -9,6 +9,7 @@ import { Container } from '@material-ui/core';
 import Card from '@material-ui/core/Card';
 import Button from '@material-ui/core/Button';
 import yogaPoses from './data/yogaPoses.json';
+import DragLogic from './components/DragLogic';
 
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 
@@ -33,60 +34,14 @@ function App() {
   const [poses, updatePoses] = useState(yogaPoses);
   const [routine, updateRoutine] = useState([]);
 
-  const handleOnDragEnd = result => {
-    // If pose is not dragged to a valid destination, keep list the same
-    if (!result.destination) return;
-
-    const source = result.source.droppableId;
-    const sourceIndex = result.source.index;
-    const destination = result.destination.droppableId;
-    const destinationIndex = result.destination.index;
-
-    if (source === destination) {
-      // Grab poses from original list
-      const items = Array.from(source === 'column-1' ? poses : routine);
-
-      // Grab moving pose & remove from list
-      const [reorderedItem] = items.splice(sourceIndex, 1);
-
-      // Add moving pose to new spot (index) in list
-      items.splice(destinationIndex, 0, reorderedItem);
-
-      //Update state
-      source === 'column-1' ? updatePoses(items) : updateRoutine(items);
-    } else if (source === 'column-1' && destination === 'column-2') {
-      // Grab poses from original lists
-      const fromItems = Array.from(poses);
-      const toItems = Array.from(routine);
-
-      // Grab moving pose & remove from list
-      const [reorderedItem] = fromItems.splice(sourceIndex, 1);
-
-      // Add moving pose to new spot (index) in list
-      toItems.splice(destinationIndex, 0, reorderedItem);
-
-      //Update state
-      updatePoses(fromItems);
-      updateRoutine(toItems);
-    } else if (source === 'column-2' && destination === 'column-1') {
-      // Grab poses from original lists
-      const fromItems = Array.from(routine);
-      const toItems = Array.from(poses);
-
-      // Grab moving pose & remove from list
-      const [reorderedItem] = fromItems.splice(sourceIndex, 1);
-
-      // Add moving pose to new spot (index) in list
-      toItems.splice(destinationIndex, 0, reorderedItem);
-
-      //Update state
-      updatePoses(toItems);
-      updateRoutine(fromItems);
-    }
-  };
-
   return (
-    <DragDropContext onDragEnd={handleOnDragEnd}>
+    <DragLogic
+      poses={poses}
+      updatePoses={updatePoses}
+      routine={routine}
+      updateRoutine={updateRoutine}
+    >
+      {/* <DragDropContext onDragEnd={handleOnDragEnd}> */}
       <Container className="landingContainer">
         <Grid
           container
@@ -99,8 +54,6 @@ function App() {
           {/* ---- LEFT COLUMN ---- */}
           <PoseBank poses={poses} />
 
-
-
           {/* ---- MIDDLE COLUMN ---- */}
           <Grid item xs={4}>
             →
@@ -112,7 +65,8 @@ function App() {
           <Button variant="outlined">Save Routine</Button>
         </Grid>
       </Container>
-    </DragDropContext>
+      {/* </DragDropContext> */}
+    </DragLogic>
   );
 }
